feat: added gRPC to wallet and users microservice and updated delete user action to check for balance before deleting

+++ ms-users/src/users/actions/delete-user.action.ts
@@ -1,11 +1,37 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Inject,
+  Injectable,
+  NotFoundException,
+  OnModuleInit,
+} from '@nestjs/common';
 import { UsersRepository } from '../repositories/transactions.repository';
+import { WalletService } from 'src/grpc/services/wallet.service';
+import type { ClientGrpc } from '@nestjs/microservices';
+import { firstValueFrom } from 'rxjs';
+import { UserWithBalanceException } from '../exceptions/user-with-balance.exception';
 
 @Injectable()
-export class DeleteUserAction {
-  constructor(private repository: UsersRepository) {}
+export class DeleteUserAction implements OnModuleInit {
+  private walletService: WalletService;
+
+  constructor(
+    private repository: UsersRepository,
+    @Inject('WALLET_PACKAGE') private client: ClientGrpc,
+  ) {}
+
+  onModuleInit() {
+    this.walletService = this.client.getService('WalletService');
+  }
 
   async execute(id: string) {
+    const balance = await firstValueFrom(
+      this.walletService.balance({ user_id: id }),
+    );
+
+    if (balance.amount > 0) {
+      throw new UserWithBalanceException(id, balance.amount);
+    }
+
     const deleted = await this.repository.delete(id);
 
     if (!deleted) {
